+++ src/ScottPlot5/ScottPlot5/Markers/FilledCircle.cs
@@ -5,7 +5,10 @@ internal class FilledCircle : IMarker
     public void Render(SKCanvas canvas, SKPaint paint, Pixel center, float size, MarkerStyle markerStyle)
     {
         float radius = size / 2;
+
         Drawing.DrawCircle(canvas, center, radius, markerStyle.FillStyle, paint);
-        Drawing.DrawCircle(canvas, center, radius, markerStyle.LineStyle, paint);
+
+        if (markerStyle.FillOutline)
+            Drawing.DrawCircle(canvas, center, radius, markerStyle.LineStyle, paint);
     }
 }