+++ src/ScottPlot5/ScottPlot5/Markers/FilledDiamond.cs
@@ -20,6 +20,8 @@ internal class FilledDiamond : IMarker
         path.AddPoly(pointsList);
 
         Drawing.DrawPath(canvas, paint, path, markerStyle.FillStyle, rect);
-        Drawing.DrawPath(canvas, paint, path, markerStyle.LineStyle);
+
+        if (markerStyle.FillOutline)
+            Drawing.DrawPath(canvas, paint, path, markerStyle.LineStyle);
     }
 }