+++ src/ScottPlot5/ScottPlot5/Markers/FilledTriangleDown.cs
@@ -22,6 +22,8 @@ internal class FilledTriangleDown : IMarker
 
         PixelRect rect = new(center, radius);
         Drawing.DrawPath(canvas, paint, path, markerStyle.FillStyle, rect);
-        Drawing.DrawPath(canvas, paint, path, markerStyle.LineStyle);
+
+        if (markerStyle.FillOutline)
+            Drawing.DrawPath(canvas, paint, path, markerStyle.LineStyle);
     }
 }