+++ src/ScottPlot5/ScottPlot5/Plottables/Scatter.cs
@@ -18,7 +18,7 @@ public class Scatter(IScatterSource data) : IPlottable, IHasLine, IHasMarker, IH
 
     public MarkerStyle MarkerStyle { get; set; } = new()
     {
-        LineWidth = 0,
+        LineWidth = 1,
         Size = 5,
         Shape = MarkerShape.FilledCircle,
     };